More fixes to the SDRAM controller.  Added pipeline to the tester to allow for better timing.  Also, shortened the length of the delay counters to improve the performance of the memory controller.

+++ rust-hdl/src/widgets/sdram/timings.rs
@@ -71,54 +71,54 @@ impl MemoryTimings {
             num_banks: 4,
         }
     }
-    pub fn t_boot(&self) -> u32 {
+    pub fn t_boot(&self) -> u16 {
         nanos_to_clocks(self.initial_delay_in_nanoseconds, self.clock_speed_hz)
     }
-    pub fn t_rp(&self) -> u32 {
+    pub fn t_rp(&self) -> u16 {
         nanos_to_clocks(self.t_rp_recharge_period_nanoseconds, self.clock_speed_hz)
     }
-    pub fn t_rfc(&self) -> u32 {
+    pub fn t_rfc(&self) -> u16 {
         nanos_to_clocks(
             self.t_rfc_autorefresh_period_nanoseconds,
             self.clock_speed_hz,
         )
     }
-    pub fn t_ras(&self) -> u32 {
+    pub fn t_ras(&self) -> u16 {
         nanos_to_clocks(
             self.t_ras_row_active_min_time_nanoseconds,
             self.clock_speed_hz,
         )
     }
-    pub fn t_rc(&self) -> u32 {
+    pub fn t_rc(&self) -> u16 {
         nanos_to_clocks(
             self.t_rc_row_to_row_min_time_nanoseconds,
             self.clock_speed_hz,
         )
     }
-    pub fn t_rcd(&self) -> u32 {
+    pub fn t_rcd(&self) -> u16 {
         nanos_to_clocks(
             self.t_rcd_row_to_column_min_time_nanoseconds,
             self.clock_speed_hz,
         )
     }
-    pub fn t_rrd(&self) -> u32 {
+    pub fn t_rrd(&self) -> u16 {
         nanos_to_clocks(
             self.t_rrd_bank_to_bank_activate_min_time_nanoseconds,
             self.clock_speed_hz,
         )
     }
-    pub fn t_wr(&self) -> u32 {
+    pub fn t_wr(&self) -> u16 {
         nanos_to_clocks(
             self.t_wr_write_recovery_time_nanoseconds,
             self.clock_speed_hz,
         )
     }
-    pub fn t_refresh_max(&self) -> u32 {
+    pub fn t_refresh_max(&self) -> u16 {
         nanos_to_clocks(self.t_refresh_max_interval_nanoseconds, self.clock_speed_hz)
     }
 }
 
-pub fn nanos_to_clocks(time_in_nanos: f64, clock_speed_hz: f64) -> u32 {
+pub fn nanos_to_clocks(time_in_nanos: f64, clock_speed_hz: f64) -> u16 {
     let clock_period_in_nanos = 1.0e9 / clock_speed_hz;
-    (time_in_nanos / clock_period_in_nanos).ceil() as u32
+    (time_in_nanos / clock_period_in_nanos).ceil() as u16
 }
